Move metaobject utils to base-services-groovy
Seems like the right place for them to be, too.
