+++ subprojects/base-services-groovy/src/main/java/org/gradle/api/internal/coerce/StringToEnumTransformer.java
