+++ subprojects/base-services-groovy/src/main/java/org/gradle/internal/metaobject/CompositeDynamicObject.java
