+++ subprojects/base-services-groovy/src/main/java/org/gradle/internal/metaobject/DynamicObjectAware.java
@@ -13,9 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.gradle.api.internal;
-
-import org.gradle.internal.metaobject.DynamicObject;
+package org.gradle.internal.metaobject;
 
 /**
  * An object that can present a dynamic view of itself.
