+++ subprojects/base-services-groovy/src/test/groovy/org/gradle/api/internal/coerce/StringToEnumTransformerTest.groovy
