+++ subprojects/base-services-groovy/src/test/groovy/org/gradle/internal/metaobject/AbstractDynamicObjectTest.java
@@ -20,7 +20,10 @@ import groovy.lang.MissingPropertyException;
 import org.junit.Test;
 
 import static org.hamcrest.Matchers.equalTo;
-import static org.junit.Assert.*;
+import static org.junit.Assert.assertFalse;
+import static org.junit.Assert.assertThat;
+import static org.junit.Assert.assertTrue;
+import static org.junit.Assert.fail;
 
 public class AbstractDynamicObjectTest {
     private final AbstractDynamicObject object = new AbstractDynamicObject() {
