+++ subprojects/base-services-groovy/src/test/groovy/org/gradle/internal/metaobject/BeanWithDynamicProperties.groovy
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.gradle.api.internal
+package org.gradle.internal.metaobject
 
 class BeanWithDynamicProperties {
     String prop
