+++ subprojects/base-services-groovy/src/test/groovy/org/gradle/internal/metaobject/CompositeDynamicObjectTest.groovy
