+++ subprojects/core/src/main/java/org/gradle/api/internal/AbstractClassGenerator.java
@@ -28,6 +28,7 @@ import org.gradle.api.NonExtensible;
 import org.gradle.api.plugins.ExtensionAware;
 import org.gradle.api.provider.HasMultipleValues;
 import org.gradle.api.provider.Property;
+import org.gradle.internal.metaobject.DynamicObjectAware;
 import org.gradle.internal.reflect.ClassDetails;
 import org.gradle.internal.reflect.ClassInspector;
 import org.gradle.internal.reflect.DirectInstantiator;
