+++ subprojects/core/src/main/java/org/gradle/api/internal/plugins/DslObject.java
@@ -17,7 +17,6 @@
 package org.gradle.api.internal.plugins;
 
 import org.gradle.api.internal.ConventionMapping;
-import org.gradle.api.internal.DynamicObjectAware;
 import org.gradle.api.internal.HasConvention;
 import org.gradle.api.internal.IConventionAware;
 import org.gradle.api.plugins.Convention;
@@ -26,6 +25,7 @@ import org.gradle.api.plugins.ExtensionContainer;
 import org.gradle.api.reflect.HasPublicType;
 import org.gradle.api.reflect.TypeOf;
 import org.gradle.internal.metaobject.DynamicObject;
+import org.gradle.internal.metaobject.DynamicObjectAware;
 
 import static org.gradle.internal.Cast.uncheckedCast;
