+++ subprojects/core/src/main/java/org/gradle/groovy/scripts/BasicScript.java
@@ -19,14 +19,14 @@ package org.gradle.groovy.scripts;
 import groovy.lang.Binding;
 import groovy.lang.MissingMethodException;
 import groovy.lang.MissingPropertyException;
-import org.gradle.api.internal.DynamicObjectAware;
-import org.gradle.api.internal.DynamicObjectUtil;
 import org.gradle.api.internal.file.HasFileOperations;
 import org.gradle.internal.logging.StandardOutputCapture;
 import org.gradle.internal.metaobject.AbstractDynamicObject;
 import org.gradle.internal.metaobject.BeanDynamicObject;
 import org.gradle.internal.metaobject.DynamicInvokeResult;
 import org.gradle.internal.metaobject.DynamicObject;
+import org.gradle.internal.metaobject.DynamicObjectAware;
+import org.gradle.internal.metaobject.DynamicObjectUtil;
 import org.gradle.internal.service.ServiceRegistry;
 
 import java.io.PrintStream;
